+++ spec/filecache_spec.rb
@@ -1,8 +1,7 @@
 require 'spec_helper'
-require 'pp'
-require_relative '../lib/jsoncache'
+require 'jsoncache'
 
-describe JSONCache do
+describe JSONCache::FileCache do
   before :all do
     @cache_directory = 'test'
     @cache = "/tmp/#{@cache_directory}"
@@ -18,27 +17,27 @@ describe JSONCache do
 
     context 'data not cached' do
       it 'should yield control if there is no valid cache' do
-        expect { |b| JSONCache.cache(@key, @options, &b) }.to(
+        expect { |b| JSONCache::FileCache.cache(@key, @options, &b) }.to(
           yield_control.once)
       end
       it 'should yield without args' do
-        expect { |b| JSONCache.cache(@key, @options, &b) }.not_to(
+        expect { |b| JSONCache::FileCache.cache(@key, @options, &b) }.not_to(
           yield_with_args)
       end
       it 'should create a cache' do
         expect(File.exist?(@cache)).to be false
-        JSONCache.cache(@key, @options) { @sample_data }
+        JSONCache::FileCache.cache(@key, @options) { @sample_data }
         expect(File.exist?(@cache)).to be true
       end
     end
 
     context 'data cached' do
       before :each do
-        @cached_data = JSONCache.cache(@key, @options) { @sample_data }
+        @cached_data = JSONCache::FileCache.cache(@key, @options) { @sample_data }
       end
 
       it 'should not yield control' do
-        expect { |b| JSONCache.cache(@key, @options, &b) }.not_to(
+        expect { |b| JSONCache::FileCache.cache(@key, @options, &b) }.not_to(
           yield_control)
       end
       it 'should return the data from the block if no cache exists' do
@@ -46,12 +45,12 @@ describe JSONCache do
       end
       it 'should return the old data if the cache is valid' do
         new_data = { 'seeya' => 'later' }
-        valid_cache_data = JSONCache.cache(@key, @options) { new_data }
+        valid_cache_data = JSONCache::FileCache.cache(@key, @options) { new_data }
         expect(valid_cache_data).to eq @cached_data
       end
       it 'should return the new data if the cache is invalid' do
         new_data = { 'seeya' => 'later' }
-        invalid_cache_data = JSONCache.cache(
+        invalid_cache_data = JSONCache::FileCache.cache(
           @key,
           @options.merge(expiry: -1)) { new_data }
         expect(invalid_cache_data).to eq new_data
